Ljung-Box

Performs the Ljung-Box test to assess the Null Hypothesis of Independence in a Time Series

diag_ljungbox(x, order = NULL, stop_lag = 20, stdres = FALSE, ...)

# S3 method for Arima
diag_ljungbox(x, stop_lag = 20, stdres = FALSE, ...)

# S3 method for default
diag_ljungbox(x, order = NULL, stop_lag = 20,
  stdres = FALSE, ...)

Arguments

x

A arima or data set.

order

A integer indicating the degrees of freedom. If `x` is not a series of residuals, then set equal to 0.

stop_lag

A integer indicating the length of lags that should be calculated.

stdres

A boolean indicating whether to standardize the residualizes (e.g. \(res/sd(res)\)) or not.